Added blob repository to the unit tests
