+++ test/UnitTest/Mocks/Repository/ServiceCollectionExtensions.cs
@@ -9,6 +9,7 @@ namespace Altinn.Platform.Storage.UnitTest.Mocks.Repository
         public static void AddMockRepositories(this IServiceCollection services)
         {
             services.AddSingleton<IDataRepository, DataRepositoryMock>();
+            services.AddSingleton<IBlobRepository, BlobRepositoryMock>();
             services.AddSingleton<IInstanceRepository, InstanceRepositoryMock>();
             services.AddSingleton<IApplicationRepository, ApplicationRepositoryMock>();
             services.AddSingleton<IInstanceEventRepository, InstanceEventRepositoryMock>();
